+++ CHANGELOG.md
@@ -2,7 +2,11 @@
 
 ## Unreleased
 
-- build(ios): Bump sentry-cocoa to 7.1.4.
+- build(ios): Bump sentry-cocoa to 7.1.4 #1700
+
+## 2.7.0-beta.1
+
+- feat: Track stalls in the JavaScript event loop as measurements #1542
 
 ## 2.6.0